% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTS.R
\name{diffM}
\alias{diffM}
\title{Difference of multivariate time series}
\usage{
diffM(zt, d = 1)
}
\arguments{
\item{zt}{A vector time series (T by k, with sample size T and dimension k)}

\item{d}{Order of differencing. Default is d=1.}
}
\value{
The differenced time series
}
\description{
Performs the difference operation of a vector time series
}
\details{
When d = 1, the command is equivalent to apply(zt,2,diff)
}
\examples{

data("mts-examples", package = "MTS")
zt <- log(qgdp[, 3:5])
xt <- diffM(zt)
}
\author{
Ruey S Tsay
}
