\name{ibmspko}
\docType{data}
\alias{ibmspko}
\title{Monthly simple returns of the stocks of 
International Business Machines (IBM) and Coca Cola (KO) and 
the S&P Composite index (SP)}
\description{
Monthly simple returns of the stocks of 
International Business Machines (IBM) and Coca Cola (KO) and 
the S&P Composite index (SP). 
The sample period is from January 1961 to December 
2011. The original data were from the Center for Research in Security Prices 
(CRSP) of the University of Chicago. The files has four columns. They 
are dates, IBM, SP, and KO.
}
\format{A 2-d list containing 612x4 observations.  The files has four columns. They 
are dates, IBM, SP, and KO.}
\source{World Almanac and Book of Facts, 1975, page 406.}
\keyword{datasets}
