% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTS.R
\name{Mtxprod}
\alias{Mtxprod}
\title{Polynomial Matrix Product}
\usage{
Mtxprod(Mtx, sMtx, p, P)
}
\arguments{
\item{Mtx}{The coefficient matrix of a regular polynomial matrix}

\item{sMtx}{The coefficient matrix of a seasonal polynomial matrix}

\item{p}{Degree of the regular polynomial matrix}

\item{P}{Degree of the seasonal polynomial matrix}
}
\value{
Coefficient matrix of the product. The product is in the form reg-AR
* sAR, etc.
}
\description{
Compute the product of two polynomial matrices
}
\author{
Ruey S. Tsay
}
