% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTS.R
\name{Mtxprod1}
\alias{Mtxprod1}
\title{Alternative Polynomial Matrix Product}
\usage{
Mtxprod1(Mtx, sMtx, p, P)
}
\arguments{
\item{Mtx}{The coefficient matrix of a regular polynomial matrix}

\item{sMtx}{The coefficient matrix of a seasonal polynomial matrix}

\item{p}{Degree of the regular polynomial matrix. p is less than P.}

\item{P}{Degree of the seasonal polynomial matrix}
}
\value{
Coefficient matrix of the product. The product matrix is in the form
sAR * reg-AR, etc.
}
\description{
Compute the product of two polynomial matrices
}
\details{
This polynomial product is used in seasonal VARMA modeling to check the
multiplicative nature between the regular and seasonal polynomial matrices
}
\author{
Ruey S. Tsay
}
