\name{qgdp}
\docType{data}
\alias{qgdp}
\title{Quarterly real gross domestic products of United Kingdom, 
Canada, and the United States}
\description{
Quarterly real gross domestic products of United Kingdom, 
Canada, and the United States from the first quarter of 1980 to the second 
quarter of 2011. The UK and CA data were originally from OECD and the US data 
from the Federal Reserve Bank of St Louis. 
}
\format{A 2-d list containing 126x5 observations. The data set consists of 5 columns: name, year, month, UK, CA, and US.}
\source{The data were downloaded from 
the FRED of the Federal Reserve Bank of St Louis. The UK data were in 
millions of chained 2006 Pounds, the CA data were in millions of chained 2002 
Canadian dollars, and the US data were in millions of chained 2005 dollars.
}

\keyword{datasets}
