% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTS.R
\name{refKronfit}
\alias{refKronfit}
\title{Refining VARMA Estimation via Kronecker Index Approach}
\usage{
refKronfit(model, thres = 1)
}
\arguments{
\item{model}{The name of a model from the command Kronfit or refKronfit}

\item{thres}{A threshold for t-ratio of individual parameter estimate. The
default is 1.}
}
\value{
Same as those of the command Kronfit.
}
\description{
This program performs model simplification of a fitted VARMA model via the
Kronecker index approach
}
\details{
For a given threshold, the program sets a parameter to zero if its t-ratio
(in modulus) is less than the threshold.
}
\references{
Tsay (2014, Chapter 4). Multivariate Time Series Analysis with R
and Financial Applications. John Wiley. Hoboken, NJ.
}
\seealso{
Kronfit
}
\author{
Ruey S. Tsay
}
