% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTS.R
\name{refVMAe}
\alias{refVMAe}
\title{Refining VMA Estimation via the Exact Likelihood Method}
\usage{
refVMAe(model, thres = 1)
}
\arguments{
\item{model}{An output object of the command VMAe or the command refVMAe
itself}

\item{thres}{A threshold for individual t-ratio of parameter estimates.
Default is 1.}
}
\value{
The same as those of the command VMAe.
}
\description{
Refines a fitted VMA model via the VMAe command by setting insignificant
parameters to zero
}
\details{
The program sets simultaneously all estimates with t-ratios less than the
threshold (in modulus) to zero.
}
\references{
Tsay (2014, Chapter 3). Multivariate Time Series Analysis with R
and Financial Applications. John Wiley. Hoboken, NJ.
}
\seealso{
VMAe, refVMA
}
\author{
Ruey S. Tsay
}
