% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varmaResiduals}
\alias{varmaResiduals}
\title{Calculate the residuals for a VARMA model}
\usage{
varmaResiduals(zt, Ph0, PH, TH, p, q, include_mean, beta)
}
\description{
Calculate the residuals for a VARMA model
}
