% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTS.R
\name{VARpsi}
\alias{VARpsi}
\title{VAR Psi-weights}
\usage{
VARpsi(Phi, lag = 5)
}
\arguments{
\item{Phi}{A k-by-kp matrix of VAR coefficients in the form Phi=[Phi1, Phi2,
..., Phip]}

\item{lag}{Number of psi-weight lags}
}
\value{
Psi-weights of a VAR model
}
\description{
Computes the psi-weight matrices of a VAR model
}
\examples{

p1 <- matrix(c(0.2, -0.6, 0.3, 1.1), 2, 2)
m1 <- VARpsi(p1, 4)
names(m1)
}
\references{
Tsay (2014, Chapter 2). Multivariate Time Series Analysis with R
and Financial Applications. John Wiley. Hoboken, NJ.
}
\author{
Ruey S. Tsay
}
