% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTS.R
\name{VARXpred}
\alias{VARXpred}
\title{VARX Model Prediction}
\usage{
VARXpred(m1, newxt = NULL, hstep = 1, orig = 0)
}
\arguments{
\item{m1}{An output object of VARX or refVARX command}

\item{newxt}{The data matrix of exogenous variables needed in forecasts.}

\item{hstep}{Forecast horizon}

\item{orig}{Forecast origin. Default is 0, meaning the last data point.}
}
\value{
Point forecasts and their standard errors
}
\description{
Computes point forecasts of a VARX model. The values of exogenous variables
must be given.
}
\details{
Uses the provided exogenous variables and the model to compute forecasts
}
\references{
Tsay (2014, Chapter 6). Multivariate Time Series Analysis with R
and Financial Applications. John Wiley. Hoboken, NJ.
}
\author{
Ruey S. Tsay
}
