% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVM.R
\name{Vech}
\alias{Vech}
\title{Half-Stacking Vector of a Symmetric Matrix}
\usage{
Vech(mtx)
}
\arguments{
\item{mtx}{A symmetric matrix}
}
\value{
a vector consisting of stacked elements of a symmetric matrix
}
\description{
Obtain the half-stacking vector of a symmetric matrix
}
\details{
Stacking a matrix into a vector using data on and below the diagonal.
}
\examples{

m1 <- matrix(c(1:9), 3, 3)
m2 <- (m1 + t(m1)) / 2
v1 <- Vech(m2)
}
\author{
Ruey S. Tsay
}
