% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTS.R
\name{comVol}
\alias{comVol}
\title{Common Volatility}
\usage{
comVol(rtn, m = 10, p = 1, stand = FALSE)
}
\arguments{
\item{rtn}{A T-by-k data matrix of k-dimensional asset returns}

\item{m}{The number of lags used to compute generalized cross-Kurtosis
matrix}

\item{p}{VAR order for the mean equation}

\item{stand}{A logical switch to standardize the returns}
}
\value{
\item{residuals}{The residuals of a VAR(p) fit}
\item{values}{Eigenvalues of the principal volatility component analysis}
\item{vectors}{Eigenvectors of the principal volatility component analysis}
\item{M}{The transformation matrix}
}
\description{
Compute the principal volatility components based on the residuals of a
VAR(p) model.
}
\details{
Perform a VAR(p) fit, if any. Then, use the residual series to perform
principal volatility component analysis. The ARCH test statistics are also
computed for the sample principal components
}
\examples{

data("mts-examples", package = "MTS")
zt <- diffM(log(qgdp[, 3:5]))
m1 <- comVol(zt, p = 2)
names(m1)
}
\references{
Tsay (2014, Chapter 7)
}
\author{
Ruey S. Tsay and Y.B. Hu
}
