% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTS.R
\name{MTSplot}
\alias{MTSplot}
\title{Multivariate Time Series Plot}
\usage{
MTSplot(data, caltime = NULL)
}
\arguments{
\item{data}{data matrix of a vector time series}

\item{caltime}{Calendar time. Default is NULL, that is, using time index}
}
\value{
Time plots of vector time series
}
\description{
Provides time plots of a vector time series
}
\details{
Provides time plots of a vector time series. The output frame depends on the
dimension of the time series
}
\examples{

xt <- matrix(rnorm(1500), 500, 3)
MTSplot(xt)
}
\author{
Ruey S. Tsay
}
