% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTS.R
\name{refECMvar}
\alias{refECMvar}
\title{Refining Error-Correction Model for VAR series}
\usage{
refECMvar(m1, thres = 1)
}
\arguments{
\item{m1}{An object of the ECMvar command or the refECMvar command}

\item{thres}{Threshold for individual t-ratio. The default is 1.}
}
\value{
Constrained estimation results of a ECM VAR model
}
\description{
Refining an estimated ECM VAR model by setting insignificant estimates to
zero
}
\details{
Set simultaneously all estimates with t-ratio less than the threshold to
zero (in modulus).
}
\references{
Tsay (2014, Chapter 5). Multivariate Time Series Analysis with R
and Financial Applications. John Wiley. Hoboken, NJ.
}
\author{
Ruey S. Tsay
}
