% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTS.R
\name{refVMA}
\alias{refVMA}
\title{Refining VMA Models}
\usage{
refVMA(model, thres = 1)
}
\arguments{
\item{model}{An output object from the command VMA or the refVMA command}

\item{thres}{A threshold for individual t-ratio of parameter estimate.
Default is 1.}
}
\value{
The same as those of the command VMA.
}
\description{
Refines a fitted VMA model by setting insignificant parameters to zero
}
\details{
The program simultaneously sets all estimates with t-ratios less than the
threshold (in modulus) to zero.
}
\references{
Tsay (2014, Chapter 3). Multivariate Time Series Analysis with R
and Financial Applications. John Wiley. Hoboken, NJ.
}
\seealso{
VMA
}
\author{
Ruey S. Tsay
}
