\name{tenstocks}
\docType{data}
\alias{tenstocks}
\title{Monthly simple returns of ten U.S. stocks}
\description{
Monthly simple returns of ten U.S. stocks. The sample period is 
from January 2001 to December 2011. Tick symbols of the ten stocks are used 
as column names for the returns. 
}
\usage{tenstocks}
\format{A 2-d list containing 132x11 observations. }
\source{The original data were from Center for 
Research in Security Prices (CRSP) of the University of Chicago. 
The first column denotes the dates.}
\keyword{datasets}